added stuff working towards getting file mode done

+++ src/utils.c
@@ -93,7 +93,7 @@ void show_cursor(void) {
 }
 
 dyn_str *dyn_str_from_string(const char *str) {
-        uint64_t len = strlen(str);
+        size_t len = strlen(str);
         char *target_str = malloc(len * sizeof(char));
         memcpy(target_str, str, len);
         dyn_str *retval = malloc(sizeof(dyn_str));
